+++ repository_manager/backend/src/routes/commit.ts
@@ -1,7 +1,7 @@
 import { prisma } from '../lib/prisma/index.js';
 import { Router } from 'express';
 import { authorizedPk } from '../middleware/auth/authHandler.js';
-import { CommittedBy } from '../lib/types/repo.js';
+import { CommitType } from '../lib/types/repo.js';
 import { commitMetrics, commitParameters, RejectedCommits } from '../lib/types/commit.js';
 import { convertCommitToNft } from '../lib/nft/nft.js';
 import { umi } from '../lib/nft/umi.js';
@@ -14,9 +14,11 @@ commitRouter.get('/', async (req, res) => {
     try {
         const commits = await prisma.commit.findMany({ where: { branchId } });
         res.status(200).json({ data: commits });
+        return;
     } catch (err) {
         console.error('Error retrieving commits:', err);
         res.status(500).send({ error: { message: 'Internal Server Error' } });
+        return;
     }
 });
 
@@ -120,11 +122,12 @@ commitRouter.get('/pending', async (req, res) => {
 commitRouter.post('/create', async (req, res) => {
     try {
         const pk = authorizedPk(res); // Get the contributor's wallet address
-        const { branchId } = req.body; // Branch ID where the commit is being made
+        const { branchId } = req; // Branch ID where the commit is being made
         // the property of the nextMergerCommit will be defined by the backend itself
         // it will not be sent in the request
+
         const {
-            committedBy,
+            commitType,
             message,
             paramHash,
             params,
@@ -134,28 +137,33 @@ commitRouter.post('/create', async (req, res) => {
             rejectedCommits,         // the rejected commits with the hash and the reject message
         }:
             {
-                committedBy: CommittedBy,
+                commitType: CommitType,
                 message: string,
                 paramHash: string,
                 params: commitParameters,
                 metrics: commitMetrics,
                 commitHash: string,
+                // reserved for the system commits
                 acceptedCommits: string[] | undefined,          // accepted and rejected commits are undefined for general commits
                 rejectedCommits: RejectedCommits | undefined,   // only needed in case of system commits 
             } = req.body;
 
         // cannot create a commit if the base model is not uploaded
-        const repo = await prisma.repository.findUnique({ where: { id: req.repoId } });        
+        const repo = await prisma.repository.findUnique({ where: { id: req.repoId } });
         if (!repo!.baseModelHash) {
             res.status(400).send({ error: { message: 'Base model not uploaded. Cannot create commit.' } });
             return;
         }
-        if (committedBy !== 'SYSTEM' && !message) {
+        if (commitType !== 'SYSTEM' && !message) {
             res.status(400).send({ error: { message: 'Commit message is required.' } });
             return;
         }
+
+        if (!branchId) {
+            throw new Error('Criticial Error: branchId not attached to response.');
+        }
         // Validate input fields
-        if (!branchId || !paramHash || !params || !metrics || !commitHash) {
+        if (!paramHash || !params || !metrics || !commitHash) {
             res.status(400).send({ error: { message: 'Complete commit information not provided.' } });
             return;
         }
@@ -210,15 +218,18 @@ commitRouter.post('/create', async (req, res) => {
             return;
         }
 
+
+        // the first commit is the only commit that is committed by the user and still a MERGER_COMMIT
+
         // Commit message section
         // Default status for a new commit is pending unless it is the first commit in the branch for which it is the merger commit
         const commitLength = await prisma.commit.count({ where: { branchId } });
-        const status = committedBy == 'SYSTEM' ? 'MERGERCOMMIT' : commitLength == 0 ? 'MERGERCOMMIT' : 'PENDING';
-        const commitMessage = committedBy == 'SYSTEM' ? '_SYSTEM_COMMIT_' : message;
+        const status = commitType == 'SYSTEM' ? 'MERGERCOMMIT' : commitLength == 0 ? 'MERGERCOMMIT' : 'PENDING';
+        const commitMessage = commitType == 'SYSTEM' ? '_SYSTEM_COMMIT_' : message;
 
 
         // if it is a merger commit
-        if (committedBy == 'SYSTEM') {
+        if (commitType == 'SYSTEM') {
             if (!acceptedCommits || !rejectedCommits) {
                 res.status(400).send({ error: { message: 'Accepted and Rejected Commits are mandatory for a merger commit.' } });
                 return;
@@ -245,13 +256,31 @@ commitRouter.post('/create', async (req, res) => {
                     include: { params: true }
                 });
 
-                // !-- this deletes the verified proof since we have already verified it
-                // !!------          change this later on need            --------------
+                // ------------------------------------ ZKML Proof handling --------------------------------------------
 
-                if (updatedCommit.params)
-                    await prisma.zKMLProof.delete({
-                        where: { paramId: updatedCommit.params.id }
-                    });
+                // !-- this deletes the verified proof since we have already verified it
+                // !!-- the ZKML proof is a big file to store. Currently we are removing it once the commits are accepted.
+                // !!------          But change this later on need            --------------
+                if (updatedCommit.params) {
+
+                    // in reality all commits must contain ZKML proofs and its verification cannot be skipped                    
+                    // the merger must check if all the ZKML proofs have been verified before creating a Merger Commit !
+
+
+                    // !! -- For Testing we consider that the ZKML proof may be empty
+                    // !! But in production environments this cannot be empty
+
+                    const proofRecordCount = await prisma.zKMLProof.count({ where: { paramId: updatedCommit.params.id } });
+                    if (proofRecordCount > 0)
+                        await prisma.zKMLProof.delete({
+                            where: { paramId: updatedCommit.params.id }
+                        });
+                    else console.warn(`High Level Warning!! Commit that does not contain ZKML proof has been merged. Commit Details: 
+                        \n Commit Hash: ${updatedCommit.commitHash}
+                        \n Committed by ID: ${updatedCommit.committerId}
+                        \n Commit Message: ${updatedCommit.message}
+                        \n Commit Time: ${updatedCommit.createdAt}`)                        
+                }
             }
 
             const rejectedCommitIds: string[] = [];
@@ -269,13 +298,18 @@ commitRouter.post('/create', async (req, res) => {
             // for the rejected commits delete all the parameters for those commits
             for (const id of rejectedCommitIds) {
                 const deletedParam = await prisma.params.delete({ where: { commitId: id } });
-                // also delete the associated zkml proof for this param
-                await prisma.zKMLProof.delete({ where: { paramId: deletedParam.id } });
+                
+                // also delete the associated zkml proof for this param if it exists 
+                
+                // For rejected commits it is allowable to have the ZKML proof as optional
+                const rejectedZKMLProofCount = await prisma.zKMLProof.count({where: {paramId: deletedParam.id}})
+                if (rejectedZKMLProofCount > 0)
+                    await prisma.zKMLProof.delete({ where: { paramId: deletedParam.id } });
             }
         }
 
         if (!params.zkmlProof) {
-            res.status(400).send({error: {message: 'Error: No ZKML proof provided as params.'}});
+            res.status(400).send({ error: { message: 'Error: No ZKML proof provided as params.' } });
             return
         }
         // extract the zkml proof for this commit
